refactor / cleanup / automation
