+++ .github/workflows/benchmarks.yml
@@ -1,0 +1,57 @@
+name: Test datagen code
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  dataset:
+    timeout-minutes: 30
+    runs-on: "ubuntu-22.04"
+    strategy:
+      matrix:
+        dataset: ["coco", "oxford_pet"]
+    steps:
+    - uses: actions/checkout@v3
+      with:
+        fetch-depth: 0
+        lfs: true
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: "3.10"
+    - uses: ./.github/workflows/build_linux_wheel
+      with:
+        python-minor-version: "10"
+    - name: Pip install
+      working-directory: python
+      run: |
+        pip3 install $(ls wheels/pylance-*-cp310*.whl)[test]
+    - name: Run test
+      working-directory: python/benchmarks
+      run: |
+        ./test_dataset.sh ${{ matrix.dataset }} 20
+  functions:
+    timeout-minutes: 30
+    runs-on: "ubuntu-22.04"
+    steps:
+    - uses: actions/checkout@v3
+      with:
+        fetch-depth: 0
+        lfs: true
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: "3.10"
+    - uses: ./.github/workflows/build_linux_wheel
+      with:
+        python-minor-version: "10"
+    - name: Pip install
+      working-directory: python
+      run: |
+        pip3 install $(ls wheels/pylance-*-cp310*.whl)[test]
+    - name: Run test
+      working-directory: python/benchmarks/functions
+      run: |
+        ./functions.py 100
