+++ python/benchmarks/bdd100k/datagen.py
@@ -1,24 +1,39 @@
 #!/usr/bin/env python
 """Parse Berkley 100K dataset."""
 
+#  Copyright (c) 2022. Lance Developers
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 import os
+import sys
 from pathlib import Path
 from typing import Union
 
 import click
 import pandas as pd
 import pyarrow as pa
-import pyarrow.fs
-from bench_utils import DatasetConverter
 
-import lance
+sys.path.append("..")
+
+from converter import DatasetConverter
 
 
 class BDD100kConverter(DatasetConverter):
     def __init__(self, uri_root: Union[str, Path]):
         super().__init__("bdd100k", uri_root)
 
-    def read_metadata(self) -> pd.DataFrame:
+    def read_metadata(self, num_rows: int = 0) -> pd.DataFrame:
         frames = []
         for split in ["train", "val"]:
             annotation = pd.read_json(
