+++ python/benchmarks/functions/README.md
@@ -1,0 +1,1 @@
+Benchmarks vectorized functions go here
