+++ python/benchmarks/oxford_pet/analytics.py
@@ -5,15 +5,16 @@ import sys
 from typing import Optional
 
 import duckdb
-import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pyarrow.compute as pc
 import pyarrow.dataset
 
+from lance.io import download_uris
+
 sys.path.append("..")
 
-from bench_utils import BenchmarkSuite, download_uris
+from suite import BenchmarkSuite, get_dataset, get_uri
 from datagen import OxfordPetConverter
 
 import lance
@@ -25,8 +26,8 @@ oxford_pet_benchmarks = BenchmarkSuite("oxford_pet")
 def label_distribution(base_uri: str, fmt: str, flavor: Optional[str]):
     if fmt == "raw":
         return get_pets_class_distribution(base_uri)
-    suffix = "" if not flavor else f"_{flavor}"
-    ds = _get_dataset(os.path.join(base_uri, f"oxford_pet{suffix}.{fmt}"), fmt)
+    uri = get_uri(base_uri, "oxford_pet", fmt, flavor)
+    ds = get_dataset(uri)
     query = "SELECT class, count(1) FROM ds GROUP BY 1"
     return duckdb.query(query).to_df()
 
@@ -35,12 +36,11 @@ def label_distribution(base_uri: str, fmt: str, flavor: Optional[str]):
 def filter_data(base_uri: str, fmt: str, flavor: Optional[str]):
     if fmt == "raw":
         return get_pets_filtered_data(base_uri)
-    suffix = "" if not flavor else f"_{flavor}"
-    uri = os.path.join(base_uri, f"oxford_pet{suffix}.{fmt}")
+    uri = get_uri(base_uri, "oxford_pet", fmt, flavor)
     if fmt == "parquet":
-        ds = _get_dataset(uri, fmt)
+        ds = get_dataset(uri)
         query = "SELECT image, class FROM ds WHERE class='pug' " "LIMIT 50 OFFSET 20"
-        return duckdb.query(query).to_df()
+        return duckdb.query(query).to_arrow_table()
     elif fmt == "lance":
         scanner = lance.dataset(uri).scanner(
             columns=["image", "class"],
@@ -48,26 +48,17 @@ def filter_data(base_uri: str, fmt: str, flavor: Optional[str]):
             limit=50,
             offset=20,
         )
-        return scanner.to_table().to_pandas()
+        return scanner.to_table()
 
 
 @oxford_pet_benchmarks.benchmark("area_histogram", key=["fmt", "flavor"])
 def compute_histogram(base_uri: str, fmt: str, flavor: Optional[str]):
     if fmt == "raw":
         return area_histogram_raw(base_uri)
-    suffix = "" if not flavor else f"_{flavor}"
-    uri = os.path.join(base_uri, f"oxford_pet{suffix}.{fmt}")
-    ds = _get_dataset(uri, fmt)
+    uri = get_uri(base_uri, "oxford_pet", fmt, flavor)
+    ds = get_dataset(uri)
     query = "SELECT histogram(size.width * size.height) FROM ds"
     return duckdb.query(query).to_df()
-
-
-def _get_dataset(uri, fmt):
-    if fmt == "parquet":
-        return pa.dataset.dataset(uri)
-    elif fmt == "lance":
-        return lance.dataset(uri)
-    raise NotImplementedError()
 
 
 def get_pets_class_distribution(base_uri):
